Py3.14/Django 6.0 support

+++ djanquiltdb/djanquiltdb/tests/__init__.py
@@ -9,16 +9,25 @@ from djanquiltdb.router import set_active_connection, DynamicDbRouter
 
 
 class CleanShardingArtifactsMixin:
-    def _fixture_setup(self):
+    @classmethod
+    def _fixture_setup(cls):
+        """
+        Save the names of the schemas that exist on start of the test case.
+        Django 6.0 calls _fixture_setup as a classmethod from _pre_setup.
+        """
+        # Can't access instance methods here, so we'll do this in setUp instead
+        super()._fixture_setup()
+    
+    def setUp(self):
         """
         Save the names of the schemas that exist on start of the test case.
         """
-        self._initial_schemas = {}
-        for db_name in self._databases_names():
-            connection_ = connections[db_name]
-            self._initial_schemas[db_name] = {s[0] for s in connection_.get_all_pg_schemas()}
-
-        super()._fixture_setup()
+        if not hasattr(self, '_initial_schemas'):
+            self._initial_schemas = {}
+            for db_name in self._databases_names():
+                connection_ = connections[db_name]
+                self._initial_schemas[db_name] = {s[0] for s in connection_.get_all_pg_schemas()}
+        super().setUp()
 
     def _post_teardown(self):
         """
@@ -26,12 +35,14 @@ class CleanShardingArtifactsMixin:
         """
         super()._post_teardown()
 
-        for db_name in self._databases_names():
-            connection_ = connections[db_name]
+        # Only clean up if _initial_schemas was initialized
+        if hasattr(self, '_initial_schemas'):
+            for db_name in self._databases_names():
+                connection_ = connections[db_name]
 
-            for schema in itertools.chain.from_iterable(connection_.get_all_pg_schemas()):
-                if schema not in self._initial_schemas[db_name]:
-                    connection_.cursor().execute('DROP SCHEMA "{}" CASCADE;'.format(schema))
+                for schema in itertools.chain.from_iterable(connection_.get_all_pg_schemas()):
+                    if schema not in self._initial_schemas[db_name]:
+                        connection_.cursor().execute('DROP SCHEMA "{}" CASCADE;'.format(schema))
 
 
 class ResetConnectionTestCaseMixin:
@@ -39,9 +50,17 @@ class ResetConnectionTestCaseMixin:
     Makes sure that at the end of each test (and as fallback, at the beginning of each test) the connection is set to
     public.
     """
-    def _pre_setup(self):
-        self._reset_connections_to_public()
+    @classmethod
+    def _pre_setup(cls):
+        # Django 6.0 calls _pre_setup as a classmethod from setUpClass
+        # We can't reset connections at class level, so this is a no-op
+        # The actual reset happens in setUp() for each test instance
         super()._pre_setup()
+
+    def setUp(self):
+        # Reset connections at the start of each test
+        self._reset_connections_to_public()
+        super().setUp()
 
     def _post_teardown(self):
         self._reset_connections_to_public()
